refactor: AccountPanelScrollPrompt
